Move telemetry and prometheus under API.
