+++ goosebit/api/telemetry/metrics.py
@@ -2,13 +2,19 @@ from opentelemetry import metrics
 from opentelemetry.sdk.metrics import MeterProvider
 from opentelemetry.sdk.resources import SERVICE_NAME, Resource
 
-from goosebit import settings
+from goosebit.settings import USERS, config
 
 from . import prometheus
 
+readers = []
+
+if config.metrics.prometheus.enable:
+    readers.append(prometheus.reader)
+
+
 resource = Resource(attributes={SERVICE_NAME: "goosebit"})
 
-provider = MeterProvider(resource=resource, metric_readers=[prometheus.reader])
+provider = MeterProvider(resource=resource, metric_readers=readers)
 metrics.set_meter_provider(provider)
 
 meter = metrics.get_meter("goosebit.meter")
@@ -25,4 +31,4 @@ users_count = meter.create_gauge(
 
 
 async def init():
-    users_count.set(len(settings.USERS))
+    users_count.set(len(USERS))
