+++ goosebit/api/telemetry/prometheus/__init__.py
@@ -1,0 +1,2 @@
+from .readers import reader  # noqa: F401
+from .routes import router  # noqa: F401
