+++ goosebit/api/telemetry/prometheus/readers.py
@@ -1,0 +1,3 @@
+from opentelemetry.exporter.prometheus import PrometheusMetricReader
+
+reader = PrometheusMetricReader()
