+++ goosebit/api/telemetry/routes.py
@@ -1,0 +1,9 @@
+from fastapi import APIRouter
+
+from goosebit.settings import config
+
+from . import prometheus
+
+router = APIRouter(prefix="/telemetry")
+if config.metrics.prometheus.enable:
+    router.include_router(prometheus.router)
