+++ goosebit/settings/schema.py
@@ -32,7 +32,6 @@ class User(BaseModel):
 
 class PrometheusSettings(BaseModel):
     enable: bool = False
-    port: int = 9090
 
 
 class MetricsSettings(BaseModel):
